requirements.txt
Pillow>=9.0.0
pillow-heif>=0.5.0; platform_system != "Windows"  # Optional but recommended for HEIC/HEIF support 